libmtd: mtd_read: Take the buffer offset into account when reading
Assuming the read() call does not return zero and the result is less
than len, the current implementation will overwrite the data already
read in buf which doesn't seem correct.

With this patch, subsequent calls to read() within the loop will now no
longer overwrite the existing contents of buf.

Signed-off-by: Marcus Prebble <marcus.prebble@axis.com>
Signed-off-by: Brian Norris <computersforpeace@gmail.com>