Makefile: install: don't look for scripts in BUILDDIR
Our ${SCRIPTS} (e.g., flash_eraseall) are not found in the build
directory; they should be found in their original location.

This fixes a typo in the Makefile refactoring, which caused 'make
install' to fail with messages like:

  make: *** No rule to make target '[...my source-build directory...]/armv7a-cros-linux-gnueabi/misc-utils/flash_eraseall'. Stop.

because the install target is looking in the wrong place for
flash_eraseall.

Fixes: 7d81790ced34 ("mtd-utils: Restructure the mtd-utils source.")
Signed-off-by: Brian Norris <computersforpeace@gmail.com>
Reviewed-by: Dongsheng Yang <yangds.fnst@cn.fujitsu.com>