Define WITHOUT_XATTR if sys/acl.h or sys/xattr.h is missing
The programs mkfs.jffs2 and mkfs.ubifs include those two headers if
WITHOUT_XATTR is not defined. Up to now, this macro is only defined
if the configure script is run with --without-xattr. If the headers
are not present on a system and the configure script is run without
special flags set, the build fails.

This patch adds a check for the presence of those headers and disables
the feature if one of the headers is missing.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>