misc-utils: initialize "ip" in docfdisk to NULL
Technically it is initializied in the for loop right before being
used. From the conditional above, we know that the for loop is
executed at least once and the variable is always initialized, but
gcc doesn't appear to perform the same reasoning.

This patch adds an initialization of the variable for the sake of
making the compiler happy.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>