nandbiterrs: Fix copy & paste fail
When porting some of the mtd-tests to user space, some code was
simplified. Among others, a while loop that iterates of page contents
was replaced with a for loop, but the old increment was left in place,
so every second byte was skipped.

This patch removes the erroneous second increment.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>