Fix unit-test header and file paths for out of tree builds
If we build mtd-utils outside the source path, we cannot use relative
paths to refere to headers in the source tree. We have to specify
absoulte paths using the top_srcdir variable.

This was done right for the utility binaries, but overlooked for the
unit test porgrams.

This patch fixes the header paths and SYSROOT variable for the unit
tests, so they build and run propperly outside the source tree.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>